Move sub-aggs into module
