+++ lib/arelastic/aggregations/date_histogram.rb
@@ -1,8 +1,15 @@
 module Arelastic
   module Aggregations
-    class DateHistogram < Arelastic::Aggregations::BucketAggregation
+    class DateHistogram < Arelastic::Aggregations::Aggregation
+      include Arelastic::Aggregations::HasSubAggregations
+
+      def initialize(name, options = {}, aggs: [])
+        super name, options
+        @aggs = aggs
+      end
+
       def as_elastic_aggregation
-        base_bucket_aggregation.merge!({'date_histogram' => options})
+        sub_aggregations.merge!({'date_histogram' => options})
       end
     end
   end
