+++ lib/arelastic/aggregations/filter.rb
@@ -1,15 +1,18 @@
 module Arelastic
   module Aggregations
-   class Filter < Arelastic::Aggregations::BucketAggregation
+   class Filter < Arelastic::Aggregations::Aggregation
+     include Arelastic::Aggregations::HasSubAggregations
+
       attr_accessor :filter
 
-      def initialize name, filter
+      def initialize(name, filter, aggs: [])
         super name
         @filter = filter
+        @aggs = aggs
       end
 
       def as_elastic_aggregation
-        base_bucket_aggregation.merge!({'filter' => convert_to_elastic(filter)})
+        sub_aggregations.merge!({'filter' => convert_to_elastic(filter)})
       end
     end
   end
