+++ lib/arelastic/aggregations/terms.rb
@@ -1,8 +1,15 @@
 module Arelastic
   module Aggregations
-    class Terms < Arelastic::Aggregations::BucketAggregation
+    class Terms < Arelastic::Aggregations::Aggregation
+      include Arelastic::Aggregations::HasSubAggregations
+
+      def initialize(name, options = {}, aggs: [])
+        super name, options
+        @aggs = aggs
+      end
+
       def as_elastic_aggregation
-        base_bucket_aggregation.merge!({'terms' => options})
+        sub_aggregations.merge!({'terms' => options})
       end
     end
   end
